long int

+++ 0x0C-more_malloc_free/101-mul.c
@@ -9,7 +9,7 @@
   */
 int main(int argc, char *argv[])
 {
-	int num1, num2, result;
+	long int num1, num2, result;
 
 	if (argc != 3)
 	{
@@ -21,7 +21,7 @@ int main(int argc, char *argv[])
 		num1 = atoi(argv[1]);
 		num2 = atoi(argv[2]);
 		result = num1 * num2;
-		printf("%d\n", result);
+		printf("%ld\n", result);
 		return (0);
 	}
 }
